requirements.txt
jira==2.0.0
python-rofi==1.0.1
requests==2.22.0
